fix: Arrowhead angle align to perfect curve
