+++ packages/math/src/curve.ts
@@ -388,6 +388,160 @@ export function curveCatmullRomCubicApproxPoints<
   return pointSets;
 }
 
+/**
+ * Fits a chord-length parameterised C2 natural cubic spline through `points`
+ * and returns one cubic Bézier segment per point pair, so the whole polyline
+ * renders as a single smooth curve that interpolates every point.
+ *
+ * The tangents `m[0..n]` at each knot are solved from a tridiagonal system
+ * (Thomas's algorithm) using the chord lengths `h[i] = |Kᵢ₊₁ − Kᵢ|` as the
+ * parameter intervals, so tightly-spaced knots don't over-influence distant
+ * ones. Bézier handles follow from the Hermite→Bézier identity:
+ *   cp1ᵢ = Kᵢ   + mᵢ   · h[i] / 3
+ *   cp2ᵢ = Kᵢ₊₁ − mᵢ₊₁ · h[i] / 3
+ *
+ * @param chordPower controls how handle length scales with chord length. At 1
+ *   handles are exactly h/3 (standard Hermite); values below 1 make short
+ *   segments curvier and long segments more taut (sub-linear scaling).
+ * @param angleCorrection in [0, 1] pulls the tangent at each interior knot
+ *   toward the chord bisector direction (`normalize(d1 + d2)`), linearly with
+ *   turn sharpness. 0 = pure C2 spline; 1 = tangent fully aligned with the
+ *   bisector. Both handles at a knot share the adjusted direction, so they
+ *   stay collinear (G1 continuity is preserved).
+ */
+export function curveChordC2Spline<Point extends GlobalPoint | LocalPoint>(
+  points: readonly Point[],
+  chordPower = 1,
+  angleCorrection = 1,
+): Curve<Point>[] {
+  const n = points.length - 1; // number of segments
+  if (n < 1) {
+    return [];
+  }
+
+  const h = new Float64Array(n);
+  for (let i = 0; i < n; i++) {
+    h[i] = Math.max(
+      1e-10,
+      Math.hypot(
+        points[i + 1][0] - points[i][0],
+        points[i + 1][1] - points[i][1],
+      ),
+    );
+  }
+
+  const mx = new Float64Array(n + 1);
+  const my = new Float64Array(n + 1);
+  const diag = new Float64Array(n + 1);
+  const rhsX = new Float64Array(n + 1);
+  const rhsY = new Float64Array(n + 1);
+
+  // Row 0 – natural BC (zero second derivative at start)
+  diag[0] = 2;
+  rhsX[0] = (3 * (points[1][0] - points[0][0])) / h[0];
+  rhsY[0] = (3 * (points[1][1] - points[0][1])) / h[0];
+
+  // Interior rows
+  for (let i = 1; i < n; i++) {
+    diag[i] = 2 * (h[i - 1] + h[i]);
+    rhsX[i] =
+      3 *
+      ((h[i] * (points[i][0] - points[i - 1][0])) / h[i - 1] +
+        (h[i - 1] * (points[i + 1][0] - points[i][0])) / h[i]);
+    rhsY[i] =
+      3 *
+      ((h[i] * (points[i][1] - points[i - 1][1])) / h[i - 1] +
+        (h[i - 1] * (points[i + 1][1] - points[i][1])) / h[i]);
+  }
+
+  // Row n – natural BC (zero second derivative at end)
+  diag[n] = 2;
+  rhsX[n] = (3 * (points[n][0] - points[n - 1][0])) / h[n - 1];
+  rhsY[n] = (3 * (points[n][1] - points[n - 1][1])) / h[n - 1];
+
+  // Forward sweep
+  for (let i = 1; i <= n; i++) {
+    const sub = i < n ? h[i] : 1;
+    const supPrev = i === 1 ? 1 : h[i - 2];
+    const w = sub / diag[i - 1];
+    diag[i] -= w * supPrev;
+    rhsX[i] -= w * rhsX[i - 1];
+    rhsY[i] -= w * rhsY[i - 1];
+  }
+
+  // Back substitution
+  mx[n] = rhsX[n] / diag[n];
+  my[n] = rhsY[n] / diag[n];
+  for (let i = n - 1; i >= 0; i--) {
+    const sup = i === 0 ? 1 : h[i - 1];
+    mx[i] = (rhsX[i] - sup * mx[i + 1]) / diag[i];
+    my[i] = (rhsY[i] - sup * my[i + 1]) / diag[i];
+  }
+
+  // Normalised tangent directions.
+  const mlen = new Float64Array(n + 1);
+  for (let i = 0; i <= n; i++) {
+    mlen[i] = Math.max(1e-10, Math.hypot(mx[i], my[i]));
+  }
+
+  // At interior knots, blend the C2 tangent direction toward the chord
+  // bisector direction by a factor proportional to turn sharpness.
+  for (let k = 1; k < n; k++) {
+    const d1x = (points[k][0] - points[k - 1][0]) / h[k - 1];
+    const d1y = (points[k][1] - points[k - 1][1]) / h[k - 1];
+    const d2x = (points[k + 1][0] - points[k][0]) / h[k];
+    const d2y = (points[k + 1][1] - points[k][1]) / h[k];
+    const dot = d1x * d2x + d1y * d2y;
+    // t: 0 = straight, 1 = hairpin
+    const t = ((1 - dot) / 2) * angleCorrection;
+    if (t < 1e-6) {
+      continue;
+    }
+    // Bisector of the two chord directions: the natural symmetric tangent.
+    const bx = d1x + d2x;
+    const by = d1y + d2y;
+    const blen = Math.hypot(bx, by);
+    if (blen < 1e-10) {
+      continue; // 180° hairpin – bisector undefined, skip
+    }
+    let px = bx / blen;
+    let py = by / blen;
+    const tx = mx[k] / mlen[k];
+    const ty = my[k] / mlen[k];
+    if (tx * px + ty * py < 0) {
+      px = -px;
+      py = -py;
+    }
+    // Linear blend of unit directions, then renormalize to preserve magnitude.
+    const blendX = tx + t * (px - tx);
+    const blendY = ty + t * (py - ty);
+    const blendLen = Math.max(1e-10, Math.hypot(blendX, blendY));
+    mx[k] = (blendX / blendLen) * mlen[k];
+    my[k] = (blendY / blendLen) * mlen[k];
+  }
+
+  const curves: Curve<Point>[] = [];
+  for (let i = 0; i < n; i++) {
+    const cpDist = Math.pow(h[i], chordPower) / 3;
+    curves.push(
+      curve(
+        pointFrom<Point>(points[i][0], points[i][1]),
+        pointFrom<Point>(
+          points[i][0] + (mx[i] / mlen[i]) * cpDist,
+          points[i][1] + (my[i] / mlen[i]) * cpDist,
+        ),
+        pointFrom<Point>(
+          points[i + 1][0] - (mx[i + 1] / mlen[i + 1]) * cpDist,
+          points[i + 1][1] - (my[i + 1] / mlen[i + 1]) * cpDist,
+        ),
+        pointFrom<Point>(points[i + 1][0], points[i + 1][1]),
+      ),
+    );
+  }
+
+  return curves;
+}
+
 export function curveOffsetPoints(
   [p0, p1, p2, p3]: Curve<GlobalPoint>,
   offset: number,
